backend/comp/requirements.txt
grpcio==1.67.1
grpcio-tools==1.67.1
sentence-transformers==3.4.1
torch==2.6.0
python-dotenv==1.0.1